Remove legacy (Win 8.1-) WinInet proxy handling
